Add date parsing and adjust CSV file and argparse
A new function 'parseDate' was added to the reports_scraper.js file to handle date parsing and timezone adjustments. Additionally, changes were made in the add_missing_info.py; the CSV file name was updated and terms of the 'speeds' dictionary were changed for accuracy. The term 'parser' used with 'argparse' was also renamed to 'arg_parser' for clarity.